refactor: scratch implementation of refactored version of sdk
