+++ Cargo.toml
@@ -10,7 +10,7 @@ members = [
     "bridge-sdk/connectors/bridge-connector-common",
     "bridge-sdk/connectors/fast-bridge",
     "bridge-sdk/connectors/omni-connector",
-    "bridge-sdk/solana-bridge-client",
+    "bridge-sdk/solana-bridge-client", "bridge-sdk/connectors/near-connector",
 ]
 
 [workspace.dependencies]
