+++ bridge-cli/src/nep141_connector_command.rs
@@ -139,7 +139,7 @@ fn nep141_connector(network: Network, cli_config: CliConfig) -> Nep141Connector 
         .eth_endpoint(combined_config.eth_rpc)
         .eth_chain_id(combined_config.eth_chain_id)
         .near_endpoint(combined_config.near_rpc)
-        .token_locker_id(combined_config.token_locker_id)
+        .token_locker_id(combined_config.near_token_locker_id)
         .bridge_token_factory_address(combined_config.bridge_token_factory_address)
         .near_light_client_address(combined_config.near_light_client_eth_address)
         .eth_private_key(combined_config.eth_private_key)
