+++ bridge-sdk/connectors/near-connector/Cargo.toml
@@ -1,9 +1,7 @@
 [package]
-name = "omni-connector"
+name = "near-connector"
 version = "0.1.0"
 edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 sha3.workspace = true
@@ -22,4 +20,3 @@ tracing.workspace = true
 eth-proof = { path = "../../eth-proof" }
 near-rpc-client = { path = "../../near-rpc-client" }
 bridge-connector-common = { path = "../bridge-connector-common" }
-near-connector = { path = "../near-connector" }
