+++ bridge-sdk/connectors/near-connector/src/lib.rs
@@ -1,0 +1,6 @@
+#[macro_use]
+extern crate derive_builder;
+
+mod near_connector;
+
+pub use near_connector::{NearConnector, NearConnectorBuilder};
