+++ bridge-sdk/connectors/omni-connector/src/lib.rs
@@ -3,4 +3,4 @@ extern crate derive_builder;
 
 mod omni_connector;
 
-pub use omni_connector::{OmniConnector, OmniConnectorBuilder};
+pub use omni_connector::{EvmConnector, EvmConnectorBuilder};
